Should be good to go after this change

+++ app/main.py
@@ -108,8 +108,7 @@ class Metadata:
                                 logger.warning(f"File {file_data['filename']} not found in {IMAGE_DIR}/uploaded")
                                 if not os.path.isfile(not_uploaded_path):
                                     logger.error(f"File {file_data['filename']} also was not found in {IMAGE_DIR}/not_uploaded")
-                                    file_data["deleted_from_meural"] = True
-                                    # raise EnvironmentError("File is missing!")
+                                    raise EnvironmentError("File is missing!")
                                 else:
                                     logger.info(f"File {file_data['filename']} found in {IMAGE_DIR}/not_uploaded - moving to proper directory")
                                     os.rename(not_uploaded_path, uploaded_path)
@@ -185,20 +184,21 @@ def delete_images_from_meural_if_needed(meural_token, icloud_album_id, album_che
             for playlist_name in playlist_names:
                 if playlist_name in Metadata.db[icloud_album_id][checksum]:
                     playlist_data = Metadata.db[icloud_album_id][checksum][playlist_name]
-                    image_filename = playlist_data['filename']
-                    try:
-                        meural.delete_image(meural_token, playlist_data['meural_id'])
-                        Metadata.mark_deleted_from_meural(icloud_album_id, checksum, playlist_name)
+                    if not playlist_data["deleted_from_meural"]:
+                        image_filename = playlist_data['filename']
+                        try:
+                            meural.delete_image(meural_token, playlist_data['meural_id'])
+                            Metadata.mark_deleted_from_meural(icloud_album_id, checksum, playlist_name)
 
-                        potential_image_location = f"{IMAGE_DIR}/not_uploaded/{image_filename}"
-                        for potential_location in (potential_image_location, potential_image_location.replace('/not_uploaded/', '/uploaded/')):
-                            if os.path.isfile(potential_location):
-                                os.remove(potential_location)
-                                logger.info(f"[✓] Successfully deleted {image_filename} from local storage")
-                                break
-                        logger.info(f"[✓] Successfully deleted {image_filename} from Meural")
-                    except Exception as e:
-                        logger.error(f"[X] Failed to delete {image_filename} from Meural: {e}")
+                            potential_image_location = f"{IMAGE_DIR}/not_uploaded/{image_filename}"
+                            for potential_location in (potential_image_location, potential_image_location.replace('/not_uploaded/', '/uploaded/')):
+                                if os.path.isfile(potential_location):
+                                    os.remove(potential_location)
+                                    logger.info(f"[✓] Successfully deleted {image_filename} from local storage")
+                                    break
+                            logger.info(f"[✓] Successfully deleted {image_filename} from Meural")
+                        except Exception as e:
+                            logger.error(f"[X] Failed to delete {image_filename} from Meural: {e}")
         else:
             logger.warning(f"Checksum {checksum} not found in metadata for album id {icloud_album_id}")
     return
